+++ userEmailDomainInfo/bin/Debug/net6.0/userEmailDomainInfo.deps.json
@@ -10,7 +10,8 @@
         "dependencies": {
           "LumenWorksCsvReader": "4.0.0",
           "Nager.PublicSuffix": "2.2.2",
-          "xunit": "2.4.1"
+          "xunit": "2.4.1",
+          "xunit.runner.visualstudio": "2.4.3"
         },
         "runtime": {
           "userEmailDomainInfo.dll": {}
@@ -754,7 +755,8 @@
             "fileVersion": "2.4.1.0"
           }
         }
-      }
+      },
+      "xunit.runner.visualstudio/2.4.3": {}
     }
   },
   "libraries": {
@@ -1336,6 +1338,13 @@
       "sha512": "sha512-7e/1jqBpcb7frLkB6XDrHCGXAbKN4Rtdb88epYxCSRQuZDRW8UtTfdTEVpdTl8s4T56e07hOBVd4G0OdCxIY2A==",
       "path": "xunit.extensibility.execution/2.4.1",
       "hashPath": "xunit.extensibility.execution.2.4.1.nupkg.sha512"
+    },
+    "xunit.runner.visualstudio/2.4.3": {
+      "type": "package",
+      "serviceable": true,
+      "sha512": "sha512-kZZSmOmKA8OBlAJaquPXnJJLM9RwQ27H7BMVqfMLUcTi9xHinWGJiWksa3D4NEtz0wZ/nxd2mogObvBgJKCRhQ==",
+      "path": "xunit.runner.visualstudio/2.4.3",
+      "hashPath": "xunit.runner.visualstudio.2.4.3.nupkg.sha512"
     }
   }
 }
